Safari-611.1.10/LayoutTests/inspector/canvas/requestContent-webgl-expected.txt

Test that CanvasAgent.requestContent can retrieve a dataURL with the current content of the WebGL canvas.


== Running test suite: Canvas.requestContentWebGL
-- Running test case: Canvas.requestContentWebGL.validCanvasId
PASS: Content should not be empty.
